requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.3; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
coloraide==2.16; python_version >= '3.8'
discord==2.3.2
discord.py==2.3.2; python_full_version >= '3.8.0'
frozenlist==1.4.1; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
multidict==6.0.5; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
